Move form content to new subdirectory
