+++ search/src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import { Route, Switch, Redirect, withRouter } from "react-router-dom"
 import {ThemeProvider} from '@material-ui/styles'
 import NavbarComponent from './Components/wrapper/navbar'
-import { Standalone as AgContextEditor } from './AgContextForm'
-import { Standalone as MetadataEditor } from './MetadataForm'
+import { Standalone as AgContextEditor } from './Components/forms/AgContextForm'
+import { Standalone as MetadataEditor } from './Components/forms/MetadataForm'
 import theme from './Components/ui/theme'
 import './App.css'
